requirements.txt
aiofiles==0.5.0
aiosqlite==0.15.0
airtable-python-wrapper==0.13.0
arrow==0.17.0
bcrypt==3.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
databases==0.3.2
decorator==4.4.2
dnspython==2.0.0
email-validator==1.1.1
fastapi==0.61.0
fastapi-users==3.0.5
h11==0.9.0
idna==2.10
ipython-genutils==0.2.0
Jinja2==2.11.2
jsonbox==1.2.1
makefun==1.9.3
MarkupSafe==1.1.1
newsapi-python==0.2.6
orjson==3.4.0
passlib==1.7.2
peewee==3.13.3
psycopg2==2.8.5
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-multipart==0.0.5
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
starlette==0.13.6
tornado==6.0.4
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.11.8
websockets==8.1
wincertstore==0.2
yarl==1.5.1
